Remove dependence on id when creating iframe links
Temporarily add `data-is-for-rex-linking="true"` attribute to uniquely
identify the element the link is generated from.

This approach assumes there is no concurrency. If there were, we would
need a better solution or a more universally unique value for the
attribute.

Use `data-is-for-rex-linking` to get `element_with_ancestors`, then remove
the attribute.

Also - make the warnings print the element instead of its id (since it
may not have one).